feat: GCN compute LDS + s_barrier + MUBUF dwordx2/x4
Extend the GCN->SPIR-V recompiler with the compute features real kernels
need beyond single-dword storage access, driven by a 6-phase cstest:

- MUBUF buffer_load/store_dwordx2/x4 (opcodes 0x0D/0x0E/0x1D/0x1E): clone
  the single-dword cases, reuse the Vdata2/Vdata4 formats; new SPIR-V
  helpers buffer_load_float2 + buffer_store_float4 (the latter reuses the
  float4 OpTypeFunction, since SPIR-V forbids duplicate type signatures).
- s_barrier (SOPP 0x0A) -> OpControlBarrier (Workgroup exec+mem,
  WorkgroupMemory|AcquireRelease). Guarded to the Linear CFG path, where a
  top-level barrier is workgroup-uniform.
- LDS: ds_write_b32/ds_read_b32 (DS 0x0D/0x36) backed by a SPIR-V
  Workgroup-storage array. shader_parse_ds's all-zero GDS guards relax to a
  per-opcode LDS branch; cs_check's lds_size!=0 EXIT is removed;
  CsStageRegisters::lds_size widened uint8->uint16 (the RSRC2 field is 9
  bits, so >=256 granules were being truncated). ShaderComputeInputInfo
  gains lds_size_dw (granule*64), sizing the array and hashed into
  ShaderGetIdCS so kernels sharing code but differing LDS size don't alias.
- Widen the MUBUF opcode decode 0x1f->0x7f so atomics (>=0x20) reach their
  KYTY_NI cases and fail loud instead of aliasing onto the dwordxN handlers.

cstest phases 1-6 verify bit-exact under strict spirv-val (direct dispatch,
load/store, async ring incl. wrap, vec4 dwordx4 copy, LDS workgroup reverse
with a real barrier). Graphics regressions (cube/gltf/instances) clean.
buffer_atomic_add deferred: the SSBO is float-typed but OpAtomicIAdd needs a
uint pointer (aliased binding) - a separable change.

+++ source/emulator/include/Emulator/Graphics/HardwareContext.h
@@ -498,7 +498,9 @@ struct CsStageRegisters
 	uint8_t  tgid_z_en      = 0;
 	uint8_t  tg_size_en     = 0;
 	uint8_t  tidig_comp_cnt = 0;
-	uint8_t  lds_size       = 0;
+	// COMPUTE_PGM_RSRC2.LDS_SIZE is a 9-bit field (mask 0x1FF); uint16_t so
+	// values >= 256 granules don't truncate.
+	uint16_t lds_size       = 0;
 };
 
 struct EsStageRegisters
